[Lesson 3] add linter, delete examples
