--- Leasson 3/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "leasson-3",
-  "version": "1.0.0",
-  "description": "",
-  "main": "test.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "node-fetch": "^3.2.10"
-  }
-}
