+++ Lesson 3/promiseToAsync.js
@@ -1,9 +1,9 @@
 // TASK: Попробуйте изменять переменные и понаблюдать, что получится.
 const secure = true;
-const sendSuccessfully = true;
+const sendSuccessfully = false;
 
 const createConnection = () => {
-  console.log('Opening connection...');
+  console.log("Opening connection...");
 
   return new Promise((resolve) => {
     // выполняем асинхронный код внутри промиса и вызываем resolve() или reject()
@@ -24,7 +24,7 @@ const createConnection = () => {
 };
 
 function prepareData(data) {
-  console.log('Preparing data...');
+  console.log("Preparing data...");
 
   return new Promise((resolve) => {
     // выполняем асинхронный код внутри промиса и вызываем resolve() или reject()
@@ -34,7 +34,7 @@ function prepareData(data) {
 }
 
 function sendData(connection, data) {
-  console.log('Sending data...');
+  console.log("Sending data...");
 
   return new Promise((resolve, reject) => {
     setTimeout(() => {
@@ -42,37 +42,38 @@ function sendData(connection, data) {
         return reject(new Error("Can't send data with insecure connection"));
       }
       if (!connection.send(data)) {
-        return reject(new Error('Error during sending'));
+        return reject(new Error("Error during sending"));
       }
       resolve(data);
     }, 2000);
   });
 }
 
-createConnection()
-  .then((connection) => {
-    const data = {
-      status: 200,
-      message: 'Hello, mister!',
-      prepared: false,
+const runFunctions = async () => {
+  try {
+    const innerFunc = async () => {
+      const connection = await createConnection();
+      const data = {
+        status: 200,
+        message: "Hello, mister!",
+        prepared: false,
+      };
+
+      const prepared = await prepareData(data, connection);
+      const serialized = await connection.serializeData(prepared);
+      const result = {
+        data: serialized,
+        connection,
+      };
+      return result;
     };
-
-    const promise = prepareData(data, connection)
-      .then((prepared) => connection.serializeData(prepared))
-      .then((serialized) => {
-        return {
-          data: serialized,
-          connection,
-        };
-      });
-
-    return promise;
-  })
-  .then((result) => {
+    const result = await innerFunc();
     const { data, connection } = result;
-    return sendData(connection, data);
-  })
-  .then(() => {
-    console.log('Sent');
-  })
-  .catch(console.error);
+    await sendData(connection, data);
+    console.log("Sent");
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+runFunctions();
